go modules, refactor, redo logging
