+++ apigee/quota/manager.go
@@ -24,7 +24,6 @@ import (
 	"github.com/apigee/istio-mixer-adapter/apigee/auth"
 	"github.com/apigee/istio-mixer-adapter/apigee/log"
 	"github.com/apigee/istio-mixer-adapter/apigee/product"
-	"istio.io/istio/mixer/pkg/adapter"
 )
 
 const (
@@ -116,8 +115,9 @@ func getQuotaID(auth *auth.Context, p *product.APIProduct) string {
 }
 
 // Apply a quota request to the local quota bucket and schedule for sync
-func (m *Manager) Apply(auth *auth.Context, p *product.APIProduct, args adapter.QuotaArgs) (*Result, error) {
+func (m *Manager) Apply(auth *auth.Context, p *product.APIProduct, args Args) (*Result, error) {
 
+	// TODO: move dup cache to mixer and create Request there to pass in
 	if result := m.dupCache.Get(args.DeduplicationID); result != nil {
 		return result, nil
 	}
